refactor: simplify settings slice.
