--- lib/features/settings/presentation/presentation.dart
@@ -1,6 +1,0 @@
-// Pages.
-export 'pages/settings_page.dart';
-
-// Widgets.
-export 'widgets/date_format_tile.dart';
-export 'widgets/theme_mode_tile.dart';
