+++ lib/features/settings/views/settings_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../presentation.dart';
+import '../widgets/date_format_tile.dart';
+import '../widgets/theme_mode_tile.dart';
 
 class SettingsPage extends StatelessWidget {
   const SettingsPage({Key? key}) : super(key: key);
